implement cooling schedules for simulated annealing + create example usage

+++ server/eas/src/algorithms.rs
@@ -8,7 +8,7 @@ pub mod simulated_annealing;
 // Shared state between all simulations. Probably the type to send to client
 #[derive(Debug)]
 pub struct SimulationState<S: SearchSpace> {
-    pub iteration: usize,
+    pub iteration: u64,
     pub current_solution: S,
     pub current_fitness: f64,
 }
